Fix unit tests, move TypographyLabelRepresentable to separate file
