+++ Tests/YMatterTypeTests/Elements/TextStyleLabelTests.swift
@@ -19,46 +19,14 @@ final class TextStyleLabelTests: XCTestCase {
         let fontSize = Constants.fontSize
         let expectedTypography = Typography.systemLabel.fontSize(fontSize)
         let expectedText = Constants.helloWorldText
-        
-        let label = TypographyLabel(typography: expectedTypography)
-        let labelConfig = { (label: TypographyLabel) in
-            label.lineBreakMode = .byTruncatingMiddle
-        }
-        labelConfig(label)
 
         // Given a TextStyleLabel with a single line of text
-        let sut = TextStyleLabel(expectedText, typography: expectedTypography)
-        
-        // we expect a value
-        XCTAssertNotNil(sut)
-        
-        // we expect the text to match the expected
-        XCTAssertEqual(sut.text, expectedText)
-        
-        // we expect the font to match the expected
-        XCTAssertEqual(sut.typography.fontSize, fontSize)
-        
-        // we expect the configuration closusure to update the label
-        XCTAssertEqual(label.lineBreakMode, .byTruncatingMiddle)
-    }
-
-    func testTypographyLabelRepresentable() throws {
-        let fontSize = Constants.fontSize
-
-        let expectedTypography = Typography.systemLabel.fontSize(fontSize)
-        let expectedText = Constants.helloWorldText
-
-        let label = TypographyLabel(typography: expectedTypography)
-        let labelConfig = { (label: TypographyLabel) in
-            label.textColor = .darkText
-        }
-        labelConfig(label)
-
-        // Given a TypographyLabelRepresentable with a single line of text
-        let sut = TextStyleLabel.TypographyLabelRepresentable(
-            text: expectedText,
+        let sut = TextStyleLabel(
+            expectedText,
             typography: expectedTypography,
-            configureTextStyleLabel: nil
+            configuration: { (label: TypographyLabel) in
+                label.lineBreakMode = .byTruncatingMiddle
+            }
         )
 
         // we expect a value
@@ -71,6 +39,41 @@ final class TextStyleLabelTests: XCTestCase {
         XCTAssertEqual(sut.typography.fontSize, fontSize)
         
         // we expect the configuration closusure to update the label
-        XCTAssertEqual(label.textColor, .darkText)
+        let labelToUpdate = TypographyLabel(typography: expectedTypography)
+        XCTAssertNotEqual(labelToUpdate.lineBreakMode, .byTruncatingMiddle)
+
+        sut.configureTextStyleLabel?(labelToUpdate)
+        XCTAssertEqual(labelToUpdate.lineBreakMode, .byTruncatingMiddle)
+    }
+
+    func testTypographyLabelRepresentable() throws {
+        let fontSize = Constants.fontSize
+
+        let expectedTypography = Typography.systemLabel.fontSize(fontSize)
+        let expectedText = Constants.helloWorldText
+
+        // Given a TypographyLabelRepresentable with a single line of text
+        let sut = TypographyLabelRepresentable(
+            text: expectedText,
+            typography: expectedTypography,
+            configureTextStyleLabel: { (label: TypographyLabel) in
+                label.textColor = .yellow
+            }
+        )
+        // we expect a value
+        XCTAssertNotNil(sut)
+        
+        // we expect the text to match the expected
+        XCTAssertEqual(sut.text, expectedText)
+        
+        // we expect the font to match the expected
+        XCTAssertEqual(sut.typography.fontSize, fontSize)
+        
+        // we expect the configuration closusure to update the label
+        let labelToUpdate = TypographyLabel(typography: expectedTypography)
+        XCTAssertNotEqual(labelToUpdate.textColor, .yellow)
+        
+        sut.configureTextStyleLabel?(labelToUpdate)
+        XCTAssertEqual(labelToUpdate.textColor, .yellow)
     }
 }
